+++ main.cjs
@@ -1,0 +1,7 @@
+'use strict';
+
+exports.load = function () {
+  console.log('[shidai3] framework extension loaded');
+};
+
+exports.unload = function () {};